pyproject.toml
[tool.poetry]
name = "fasthtml-market"
version = "0.1.0"
description = ""
authors = ["Pol Alvarez <pol.avms@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-fasthtml = "^0.2.4"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
